Move SectorAllocator into pkg/filesystem/pool
SectorAllocator is only used by the FilePool code, so there's no point
in keeping it in pkg/filesystem.
