+++ pkg/filesystem/pool/bitmap_sector_allocator.go
@@ -1,4 +1,4 @@
-package filesystem
+package pool
 
 import (
 	"fmt"
