+++ pkg/filesystem/pool/block_device_backed_file_pool.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"io"
 
-	re_filesystem "github.com/buildbarn/bb-remote-execution/pkg/filesystem"
 	"github.com/buildbarn/bb-storage/pkg/blockdevice"
 	"github.com/buildbarn/bb-storage/pkg/filesystem"
 
@@ -14,7 +13,7 @@ import (
 
 type blockDeviceBackedFilePool struct {
 	blockDevice     blockdevice.BlockDevice
-	sectorAllocator re_filesystem.SectorAllocator
+	sectorAllocator SectorAllocator
 	sectorSizeBytes int
 	zeroSector      []byte
 }
@@ -24,7 +23,7 @@ type blockDeviceBackedFilePool struct {
 // device tends to be faster than using a directory on a file system,
 // for the reason that no metadata (e.g., a directory hierarchy and
 // inode attributes) needs to be stored.
-func NewBlockDeviceBackedFilePool(blockDevice blockdevice.BlockDevice, sectorAllocator re_filesystem.SectorAllocator, sectorSizeBytes int) FilePool {
+func NewBlockDeviceBackedFilePool(blockDevice blockdevice.BlockDevice, sectorAllocator SectorAllocator, sectorSizeBytes int) FilePool {
 	return &blockDeviceBackedFilePool{
 		blockDevice:     blockDevice,
 		sectorAllocator: sectorAllocator,
